addAbkineticsModel

This function adds an antibody kinetics model to the model definition.

Usage

addAbkineticsModel(id, biomarker, exposureType, pars, funcForm)

Arguments

biomarker

The name of the biomarker.

pars

The parameters of the model.

funcForm

The antibody kinetics function.

is

The name of the biomarker.

exposureName

The name of the exposure type.

Value

A list with the biomarker name, the exposure name, whether the exposure is inferred, the parameters and the antibody kinetics function.