addAbkineticsModelHier

This function adds an antibody kinetics model to the model definition.

addAbkineticsModelHier(
  id,
  biomarker,
  exposureType,
  pars,
  parsHier,
  dataHier,
  funcForm
)

Arguments

id

The name of the biomarker.

biomarker

The name of the biomarker.

exposureType

The name of the exposure type.

pars

The parameters of the model.

parsHier

The parameters of the Hierarchical model.

dataHier

The data for the Hierarchical model.

funcForm

The antibody kinetics function.

Value

A list with the biomarker name, the exposure name, whether the exposure is inferred, the parameters and the antibody kinetics function.